Cleaned up Code a bit
Working on AI for white blood cells and a function that will hande the rotation for all the cells in the game, using rotate acceleration and velocity to smooth out turns. Also I am taking a break from the flagellum object as I don't know if I want to use a sprite instead or not. ;)